pyproject.toml
[tool.poetry]
name = "dirtty"
version = "0.1.0"
description = "Terminal-based AI assistant for querying your local document collections"
authors = ["Levi Smith <levi@fynx.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
textual = "^0.76.0"
llama-index = "^0.10.67.post1"
pypdf = "*"
langchain = "*"
ollama = "*"
pyyaml = "^6.0"
llama-index-llms-ollama = "^0.2.2"
llama-index-embeddings-huggingface = "^0.2.3"
humanize = "^4.6.0"
click = "^8.1.6"
xdg-base-dirs = "^6.0.1"
aiosqlite = "^0.20.0"
click-default-group = "^1.2.4"
greenlet = "^3.0.3"
pyperclip = "^1.8.2"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
mypy = "^1.3.0"
pre-commit = "^3.3.2"
textual-dev = "^1.0.1"
pyinstrument = "^4.6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dirtty = "dirtty.main:main"

[tool.mypy]
ignore_missing_imports = true